# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AskHealth AI Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /communities/{communityId}/posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts in a community
      parameters:
        - name: communityId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post in a community
      parameters:
        - name: communityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /communities/{communityId}/posts/pinned:
    get:
      operationId: getPinnedPosts
      tags: [posts]
      summary: Get pinned posts in a community
      parameters:
        - name: communityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pinned posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post by ID
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /posts/{postId}/pin:
    patch:
      operationId: pinPost
      tags: [posts]
      summary: Pin/unpin a post (admin/moderator only)
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinPostBody"
      responses:
        "200":
          description: Updated pin state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}/upvote:
    post:
      operationId: upvotePost
      tags: [posts]
      summary: Upvote a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Upvote result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpvoteResult"

  # ─── Comments ──────────────────────────────────────────────────────────
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current authenticated user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}/ban:
    patch:
      operationId: banUser
      tags: [users]
      summary: Ban/unban a user (admin/moderator only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── Gamification ────────────────────────────────────────────────────────
  /communities/{communityId}/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [gamification]
      summary: Get community leaderboard (weekly and all-time)
      parameters:
        - name: communityId
          in: path
          required: true
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
            enum: [weekly, all_time]
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /users/{userId}/achievements:
    get:
      operationId: getUserAchievements
      tags: [gamification]
      summary: Get a user's badges and achievements
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Achievement"
...
components:
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
        communityId:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        authorLevel:
          type: integer
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        isPinned:
          type: boolean
        upvoteCount:
          type: integer
        commentCount:
          type: integer
        hasUpvoted:
          type: boolean
        isBroadcast:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - communityId
        - authorId
        - authorName
        - title
        - content
        - isPinned
        - upvoteCount
        - commentCount
        - hasUpvoted
        - isBroadcast
        - createdAt
        - updatedAt
    CreatePostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
      required:
        - title
        - content
    UpdatePostBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    PinPostBody:
      type: object
      properties:
        isPinned:
          type: boolean
      required:
        - isPinned
    UpvoteResult:
      type: object
      properties:
        upvoteCount:
          type: integer
        hasUpvoted:
          type: boolean
      required:
        - upvoteCount
        - hasUpvoted
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        displayName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, moderator, member]
        isBanned:
          type: boolean
        healthCredits:
          type: integer
        level:
          type: integer
        weeklyCredits:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - displayName
        - email
        - role
        - isBanned
        - healthCredits
        - level
        - weeklyCredits
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, moderator, member]
      required:
        - role
    BanUserBody:
      type: object
      properties:
        isBanned:
          type: boolean
      required:
        - isBanned
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        level:
          type: integer
        credits:
          type: integer
        badgeCount:
          type: integer
      required:
        - rank
        - userId
        - displayName
        - level
        - credits
        - badgeCount
    Achievement:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        badgeName:
          type: string
        badgeDescription:
          type: string
        badgeIcon:
          type: string
        earnedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - badgeName
        - badgeDescription
        - badgeIcon
        - earnedAt
...